[FIX,TEST] Re-land CPP-043 and cast the cluster bin size
CPP-043: the second isotopic peak of the cross-link generator's precursor
peaks was placed at the charged mass plus the isotope spacing divided by the
charge, then written as an m/z. mono_pos is a charged mass, so the spacing
belongs inside the division: at charge 2 and above the companion landed far
from its precursor, and only charge 1 was right. The intact, water-loss and
ammonia-loss precursors all had it. The fix shipped in core-v4.0.0-ci.3 and
was reverted in ci.4 together with CPP-042 to restore the OpenPepXL reference
outputs; CPP-042 returned in ci.7 and this completes that pair. The new test
checks, for charges 1 to 3, that each precursor peak and its companion differ
by one isotope spacing of that charge.

ClusterHierarchical::cluster now casts the bin size to float where it builds
each BinnedSpectrum. MSVC reported the narrowing from inside <xutility> in
every consumer that instantiates the function, which is why FLASH carries a
/wd4244 suppression; that can go once this ships.

+++ src/openms/include/OpenMS/ML/CLUSTERING/ClusterHierarchical.h
@@ -142,7 +142,9 @@ public:
     for (Size i = 0; i < data.size(); i++)
     {
       // double sz(2), UInt sp(1);
-      binned_data.emplace_back(data[i], sz, false, sp, offset);
+      // BinnedSpectrum takes the bin size as float; cast here, or MSVC reports the narrowing
+      // from inside <xutility> in every consumer that instantiates this function.
+      binned_data.emplace_back(data[i], static_cast<float>(sz), false, sp, offset);
     }
 
     // create distancematrix for data with comparator
